+++ src/actions/WorkflowActions.js
@@ -5,7 +5,9 @@ import {
   UPDATE_WORKFLOW,
   REMOVE_WORKFLOW,
   OPEN_WORKFLOW_MODAL,
-  CLOSE_WORKFLOW_MODAL
+  CLOSE_WORKFLOW_MODAL,
+
+  NO_AUTH_ERROR
 } from './types'
 
 import { apiError } from './Errors'
@@ -13,6 +15,9 @@ import { apiError } from './Errors'
 import { ROOT_URL } from './config'
 
 export const fetchWorkflows = () => {
+  if (!window.localStorage.getItem('token')) {
+    return dispatch => dispatch({ type: NO_AUTH_ERROR })
+  }
   return (dispatch) => {
     axios.get(`${ROOT_URL}/workflow`)
       .then(response => fetchWorkflowsSuccess(dispatch, response))
@@ -28,6 +33,9 @@ const fetchWorkflowsSuccess = (dispatch, response) => {
 }
 
 export const addWorkflow = (props) => {
+  if (!window.localStorage.getItem('token')) {
+    return dispatch => dispatch({ type: NO_AUTH_ERROR })
+  }
   return (dispatch) => {
     axios.post(`${ROOT_URL}/workflow`, props)
       .then(response => addWorkflowSuccess(dispatch, response))
@@ -44,6 +52,9 @@ const addWorkflowSuccess = (dispatch, response) => {
 }
 
 export const updateWorkflow = (entity) => {
+  if (!window.localStorage.getItem('token')) {
+    return dispatch => dispatch({ type: NO_AUTH_ERROR })
+  }
   return (dispatch) => {
     axios.put(entity._links.self.href, entity)
       .then(response => updateWorkflowSuccess(dispatch, response))
@@ -60,6 +71,9 @@ const updateWorkflowSuccess = (dispatch, response) => {
 }
 
 export const removeWorkflow = (entity) => {
+  if (!window.localStorage.getItem('token')) {
+    return dispatch => dispatch({ type: NO_AUTH_ERROR })
+  }
   return (dispatch) => {
     axios.delete(entity._links.self.href)
       .then(() => removeWorkflowSuccess(dispatch, entity))
